Fix Non-ASCII Characters and Improve GTK Unicode Support
Replace all non-ASCII characters (em dashes, curly quotes, and special
bullets) with standard ASCII equivalents throughout the codebase to
ensure proper work of `clangd` on GTK.

Add explicit dependency on `fonts-noto` package in build documentation
to ensure proper Unicode symbol rendering in the editor on GTK builds.

Add new Python utility script (check-for-non-ascii.py) to detect
non-ASCII characters in source files during development, with optional
verbose mode to show full context.

Fix type cast warning in GitConsole by explicitly casting GetItemCount()
return value to size_t.

Extend LLM event system with new user reply timeout and error events,
allowing the UI to properly dismiss info bars when user prompts fail
or timeout.

Update AI documentation prompt generation to explicitly request ASCII-only
output, preventing non-ASCII characters in generated documentation comments.

Initialize clEditEventsHandler for both input and output text controls
in ChatAIWindow to ensure consistent editing behavior.

+++ git/GitConsole.cpp
@@ -153,7 +153,7 @@ wxString GenerateRandomFile()
 
 void DeleteAllItems(wxDataViewListCtrl* list)
 {
-    for (size_t i = 0; i < list->GetItemCount(); ++i) {
+    for (size_t i = 0; i < static_cast<size_t>(list->GetItemCount()); ++i) {
         auto item_data = list->GetItemData(list->RowToItem(i));
         GitClientData* cd = reinterpret_cast<GitClientData*>(item_data);
         wxDELETE(cd);
@@ -1121,7 +1121,7 @@ void GitConsole::FinaliseReleaseNotes(const wxString& complete_response)
 std::shared_ptr<std::string> GitConsole::AllocateBuffer()
 {
     auto buffer = std::make_shared<std::string>();
-    buffer->reserve(32 * 1024); // assumes ~32KB worst‑case
+    buffer->reserve(32 * 1024); // assumes ~32KB worst case
     return buffer;
 }
